refactor: delete dead code in storage
Remove the disabled admission plugin package, the unused aggregated-data
computation (ComputeAggregatedData and the ContainerProfileProcessor
getAggregatedData method, neither with a production caller), the uncalled
deleteWrongSchemaVersion cleanup path, and the never-assigned
SharedInformerFactory option field. Preserve the shared fakeStorage test
helper the aggregator test carried.
