--- pkg/admission/wardleinitializer/wardleinitializer_test.go
@@ -1,59 +1,0 @@
-/*
-Copyright 2017 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package wardleinitializer_test
-
-import (
-	"context"
-	"testing"
-	"time"
-
-	"github.com/kubescape/storage/pkg/admission/wardleinitializer"
-	"github.com/kubescape/storage/pkg/generated/clientset/versioned/fake"
-	informers "github.com/kubescape/storage/pkg/generated/informers/externalversions"
-	"k8s.io/apiserver/pkg/admission"
-)
-
-// TestWantsInternalWardleInformerFactory ensures that the informer factory is injected
-// when the WantsInternalWardleInformerFactory interface is implemented by a plugin.
-func TestWantsInternalWardleInformerFactory(t *testing.T) {
-	cs := &fake.Clientset{}
-	sf := informers.NewSharedInformerFactory(cs, time.Duration(1)*time.Second)
-	target := wardleinitializer.New(sf)
-
-	wantWardleInformerFactory := &wantInternalWardleInformerFactory{}
-	target.Initialize(wantWardleInformerFactory)
-	if wantWardleInformerFactory.sf != sf {
-		t.Errorf("expected informer factory to be initialized")
-	}
-}
-
-// wantInternalWardleInformerFactory is a test stub that fulfills the WantsInternalWardleInformerFactory interface
-type wantInternalWardleInformerFactory struct {
-	sf informers.SharedInformerFactory
-}
-
-func (f *wantInternalWardleInformerFactory) SetInternalWardleInformerFactory(sf informers.SharedInformerFactory) {
-	f.sf = sf
-}
-func (f *wantInternalWardleInformerFactory) Admit(ctx context.Context, a admission.Attributes, o admission.ObjectInterfaces) error {
-	return nil
-}
-func (f *wantInternalWardleInformerFactory) Handles(o admission.Operation) bool { return false }
-func (f *wantInternalWardleInformerFactory) ValidateInitialization() error      { return nil }
-
-var _ admission.Interface = &wantInternalWardleInformerFactory{}
-var _ wardleinitializer.WantsInternalWardleInformerFactory = &wantInternalWardleInformerFactory{}
