+++ pkg/registry/file/containerprofile_processor.go
@@ -2,11 +2,8 @@ package file
 
 import (
 	"context"
-	"crypto/sha256"
-	"encoding/hex"
 	"errors"
 	"fmt"
-	"sort"
 	"strconv"
 	"strings"
 	"time"
@@ -770,64 +767,6 @@ func (a *ContainerProfileProcessor) updateProfileStatus(ctx context.Context, key
 	}
 
 	return newTimeSeries, false, nil
-}
-
-// getAggregatedData computes various data of the aggregated profile.
-// A profile status is completed only if all its main containers are completed.
-// A profile completion is full only if all its init/main containers are full.
-// A profile sync checksum is the checksum of all container checksums.
-func (a *ContainerProfileProcessor) getAggregatedData(ctx context.Context, key string, parts map[string]string) (string, string, string) {
-	mainContainers := 0
-	completed := 0
-	full := 0
-	var tooLarge bool
-	status := helpers.Learning
-	completion := helpers.Partial
-	hasher := sha256.New()
-	// Sort keys to ensure deterministic iteration order for consistent checksum
-	keys := make([]string, 0, len(parts))
-	for k := range parts {
-		keys = append(keys, k)
-	}
-	sort.Strings(keys)
-	for _, key := range keys {
-		cpCtx, cpCancel := context.WithTimeout(ctx, 5*time.Second)
-		defer cpCancel()
-		profile, err := a.ContainerProfileStorage.GetContainerProfileMetadata(cpCtx, key)
-		if err != nil {
-			logger.L().Debug("ContainerProfileProcessor.getAggregatedData - failed to get profile", loggerhelpers.Error(err), loggerhelpers.String("key", key))
-			continue
-		}
-		// only main containers are considered for aggregated status
-		if profile.Annotations[helpers.ContainerTypeMetadataKey] == "containers" {
-			mainContainers++
-			if profile.Annotations[helpers.StatusMetadataKey] == helpers.Completed {
-				completed++
-			}
-		}
-		if profile.Annotations[helpers.CompletionMetadataKey] == helpers.Full {
-			full++
-		}
-		if profile.Annotations[helpers.StatusMetadataKey] == helpers.TooLarge {
-			tooLarge = true
-		}
-		checksum := profile.Annotations[helpers.SyncChecksumMetadataKey]
-		parts[key] = checksum
-		hasher.Write([]byte(checksum)) // profile.Parts is sorted so the checksum is consistent
-	}
-	if completed == mainContainers && mainContainers > 0 {
-		status = helpers.Completed
-	} else if tooLarge {
-		status = helpers.TooLarge
-	}
-	if full == len(parts) {
-		completion = helpers.Full
-	}
-	hash := hex.EncodeToString(hasher.Sum(nil))
-	logger.L().Debug("ContainerProfileProcessor.getAggregatedData - returning", loggerhelpers.String("key", key),
-		loggerhelpers.Int("mainContainers", mainContainers), loggerhelpers.Int("completed", completed), loggerhelpers.Int("full", full),
-		loggerhelpers.String("status", status), loggerhelpers.String("completion", completion), loggerhelpers.String("hash", hash))
-	return status, completion, hash
 }
 
 func DeflateContainerProfileSpec(container softwarecomposition.ContainerProfileSpec, sbomSet mapset.Set[string], settings dynamicpathdetector.CollapseSettings) softwarecomposition.ContainerProfileSpec {
